Feat: Add redis_client object
To fix the circular import i needed to remove the contents of __init__.py and move it over to the app.py as its still persisting to enter a circular import.

Now that we have fixed the circular import we can initialize the redis_client inside the create_app method.

Files:
- app/__init__.py
- app/app.py