Feat: Add Task class
Now as to have a simple idea of what to store onto the redis storage I need to have it stored on the program for a bit till it gets stored on the redis.

Where the Task class comes into play, it holds the attributes of the task and simple functionality to convert to a dictionary and load the attributes from a dictionary.

As for the __init__.py it contains the import for ease of access rather than importing from every file.

Files:
- app/models/__init__.py
- app/models/task.py