Feat: Add task services
Now since we have the task model in place we need to have the task services in place such as creating a task, looking for a task, deleting a task, and updating the task.

The TaskService class contains these functionality which will later be used in the controllers & routers.

The __init__.py file contains the import for ease of access.

Files:
- app/services/__init__.py
- app/services/task_service.py