Feat: Task endpoints work
After testing a lot and debugging, the task endpoints are working and giving the desired status codes.

The changes that were needed were to ensure adding json.dumps while adding to redis.

Now I need to stylize the directory inside of redis but that would be later, I also am thinking of having it a couple of keys under the ID rather than it being a json as to be able to edit the specifics easier.

Files:
- app/app.py
- app/controllers/task_controller.py
- app/services/task_service.py
- app/utils/redis_client.py