Feat: Add user service logic
Now that the model is in place we can add the business logic of the API into our program. The UserService class contains the basic functions like create user, get user, update user, and delete user.

Also added the import into the __init__.py for ease of access to the class.

Files:
- app/services/__init__.py
- app/services/user_service.py