feat: Tasks in Projects
Now we can assign tasks to certain projects and retrieve all the tasks in a certain project.

PUT -> /api/projects/<project_id>/<task_id>
We assign the task id into a list of tasks in the project using this PUT request

GET -> /api/projects/<project_id>/tasks
We retrieve all the tasks that are inside the list of tasks of a certain project using this GET request.

Files:
- app/controllers/project_controller.py (modified)
- app/models/projects.py (modified)
- app/routes/projects.py (modified)
- app/services/project_service.py (modified)