feat: Retrieve user tasks
With this feature we can retrieve all the tasks that a user is assigned to. Making it easier for a user to know what is assigned to him/her and what isn't.

GET -> /api/users/<user_id>/tasks
Retrieve all the tasks that are assigned to the entered user_id.

Files:
- app/controllers/user_controller.py (modified)
- app/routes/users.py (modified)
- app/services/user_service.py (modified)